Remove the Cactus engine, bindings, and bundled model
Delete the :cactus and :cactus-native modules (the prebuilt
libcactus_engine.so, the Kotlin bindings and cinterop definitions, the
CPU probe shim, and the iOS static archives), the bundled
needle-pebble-ft archive and its APK asset symlink, and the Cactus-era
provider stack (CactusModelProvider, ModelZipInstaller,
CactusModelPins) with its four test suites; the whisper catalog,
installer, and provider landed with equivalent coverage in the
preceding commits. Nothing on the runtime or test classpath consumed
any of it after the transcription swap.

Ripples: the cactusProKey property and the four CACTUS_* build-config
fields are gone (the iOS model downloader writes the catalog
generation string as its marker instead); :libindex loses a declared
:cactus dependency its sources never imported; androidApp drops the
/ios/** packaging exclude that existed only to keep the deleted iOS
archives out of the APK; stale engine references in comments are
reworded. CactusSTTMode and promoteSingleRootDir stay, with a comment
recording why: the enum's entry names are persisted inside the
serialized config on every existing install, and the iOS downloader
still calls the directory promotion helper.

A classpath probe (CactusAbsenceTest, same shape as the Mixpanel
probe) fails the suite if the binding classes ever return, whatever an
upstream merge does to the source tree.

Verified: full unit sweep green including the new absence probe;
androidApp assembles for debug including the androidTest APK; the
debug APK contains libwhispercpu.so and libwhisperjni.so and no cactus
library, no bundled model assets.
