--- composeApp/src/androidMain/assets/models/needle-pebble-ft-cq4.zip
@@ -1,1 +1,0 @@
-../../../../../models/needle-pebble-ft-cq4.zip
