+++ util/src/iosMain/kotlin/coredevices/util/models/ModelDownloadManager.ios.kt
@@ -1,7 +1,6 @@
 package coredevices.util.models
 
 import co.touchlab.kermit.Logger
-import coredevices.util.CommonBuildKonfig
 import kotlinx.cinterop.BetaInteropApi
 import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -143,7 +142,7 @@ private class DownloadDelegate(private val manager: ModelDownloadManager) : NSOb
             val outputDirIo = IoPath(outputDir.toString())
             promoteSingleRootDir(outputDirIo)
             SystemFileSystem.sink(IoPath(outputDirIo, ".cactus_version")).buffered().use {
-                it.write(CommonBuildKonfig.CACTUS_WEIGHTS_VERSION.encodeToByteArray())
+                it.write(WhisperModelCatalog.GENERATION.encodeToByteArray())
             }
             logger.i {"Model $slug downloaded and extracted to $outputDir"}
             manager.updateDownloadStatus(ModelDownloadStatus.Idle)
